requirements.txt
gitpython==3.1.31
pandas==1.4.1
scipy==1.10.1